Added option to not mmap
Added option in constructor to not mmap shm. reads are on the order of 2*10^-6 and 10^-5 seconds for mmapped and not mmapped. Writes are on the order of 10^-5 and 5*10^-5 for mmapped and not.

Also got rid of last access time (unnecessary so got rid of it to speed up reads and writes)